segunda

+++ routes/api.php
@@ -43,6 +43,6 @@ Route::namespace('Api')->group(function () {
     Route::get('/users', 'UsersController@index');
     Route::get('/users/{user}', 'UsersController@show');
     Route::put('/users/{user}', 'UsersController@update');
-    Route::delete('/users/{user}', 'UsersController@destroy');
+    Route::delete('/users/{user}', 'UsersController@destroy')->middleware('auth:api');
     Route::post('/users', 'UsersController@store');
 });
